# Chore: Only instantiate fileEntryCache when cache flage set (perf)