--- a/lib/cli-engine.js
+++ b/lib/cli-engine.js
@@ -398,15 +398,17 @@ class CLIEngine {
         this.options = options;
         this.linter = new Linter();
 
-        const cacheFile = getCacheFile(this.options.cacheLocation || this.options.cacheFile, this.options.cwd);
+        if (options.cache) {
+            const cacheFile = getCacheFile(this.options.cacheLocation || this.options.cacheFile, this.options.cwd);
 
-        /**
-         * Cache used to avoid operating on files that haven't changed since the
-         * last successful execution (e.g., file passed linting with no errors and
-         * no warnings).
-         * @type {Object}
-         */
-        this._fileCache = fileEntryCache.create(cacheFile);
+            /**
+             * Cache used to avoid operating on files that haven't changed since the
+             * last successful execution (e.g., file passed linting with no errors and
+             * no warnings).
+             * @type {Object}
+             */
+            this._fileCache = fileEntryCache.create(cacheFile);
+        }
 
         // load in additional rules
         if (this.options.rulePaths) {
@@ -495,6 +497,11 @@ class CLIEngine {
             fileCache = this._fileCache,
             configHelper = this.config;
         let prevConfig; // the previous configuration used
+        const cacheFile = getCacheFile(this.options.cacheLocation || this.options.cacheFile, this.options.cwd);
+
+        if (!options.cache && fs.existsSync(cacheFile)) {
+            fs.unlinkSync(cacheFile);
+        }
 
         /**
          * Calculates the hash of the config file used to validate a given file
@@ -570,8 +577,6 @@ class CLIEngine {
                     // move to the next file
                     return;
                 }
-            } else {
-                fileCache.destroy();
             }
 
             debug(`Processing ${filename}`);